Split commands route into three focused files to reduce complexity
commands.ts now owns only the GET list handler; commandMutations.ts
handles POST add/update/remove; commandAssignments.ts handles POST
assign/unassign. commands.ts re-exports a combined router so server.ts
and the test file are unchanged.
